Disable password reset except on Endless hardware
The password reset feature is only desired on Endless hardware, not on
the free downloadable images nor on OEM images. Image type is determined
by checking the eos-image-version xattr of /sysroot. A gsetting is
provided to override the default behavior. (Note that the gsetting has
to be set for the gdm user, not your own user account.)

https://phabricator.endlessm.com/T13460