Do not show Builder in places like the alt-tab dialog
When launching Builder hint the window tracker about
it and track the appearance of the next Builder
window. Store the association of the app window
and the builder window.

In the windowManager store as well that Builder
window. When the Builder window is destroyed clean
up the association in the shell-window-tracker.