Map Builder window to be another App window
This handles the Builder window as another App
window. It will be displayed in the navigation
bar and alt-tab overlay like would be done for
multiple windows of the same App. It allows
as well to select one of those windows from the
extended selection options and animates in
certain cases (when the other coding session
window is the currently focused one).

Builder starts maximized. We have to unmaximize
first in order to be able to resize the window
in the cases where the App window is not
maximized.