icon grid: Allow icons to cover parts of the icon shadow
The overridable createIcon function can now return an array of
icons. If the "shadowAbove" setting is set to true, then the first
icon in the array will be below the shadow and the rest will be above.

The reason for this change is to allow the badge to be on top of the
shadow.