Fix the first animation and add more handling to the restacking handler
Only the animation itself has to wait for the first frame of
the actor. Move the preparation code (like resizing) out of
that case.

Make sure the that the restacking does not trigger another
animation. Also add an early check if it is the same window
that is meant to get into foreground and that we already
handled.