Synchronize shell startup
The asynchronous nature of extension loading, session loading, and more,
makes the code racy as to what is initialized first, and hard to debug.
Additionally, since gjs is single-threaded, the only code we're running
in a thread anyway is readdir, which is going to be I/O bound, so the
code here is actually likely to be faster.

Drop this in favor of some good old fashioned synchronous loading.

Cherry picked from upstream commit
da4238ec681c7ca5f682867933e1f8fcbc623bf8. Only the changes in
js/extensionPrefs/main.js were considered.