tests: Adapt to new GjsCoverage API
In the upcoming version of GJS, the gjs/coverage.h header disappears, and
GjsCoverage is available through the public API in gjs.h.

The constructor gjs_coverage_new_for_cache() disappears as well; the only
public constructor is gjs_coverage_new(), which takes the output
directory that was previously passed to gjs_coverage_write_statistics().

https://phabricator.endlessm.com/T13622