network: unset _activeNetwork when access point changes
The current code expects NMDevice to always trigger the
access-point-removed signal before notify::active-access-point. I
haven't done enough testing to be confident that that's a bad
assumption, but at least it doesn't seem very safe. Let's become more
robust to the order that these signals are emitted in. This is a
speculative fix for a JS error that occurs when
NMDeviceWireless._createActiveConnectionItem is called from
NMDeviceWireless._createSection when this._activeNetwork points to
some stale network and there is not actually any active network or
access point anymore. This is matches the behavior of the current
upstream code.

https://phabricator.endlessm.com/T4444