network: Create menu item only if active AP exists
If there's no active AP then a call to _createActiveConnectionItem is
just going to cause a JS error. We have to handle it here rather than
inside _createActiveConnectionItem because _activeConnectionItem should
always be set after a call to _createActiveConnectionItem, and we have
to handle the case where it's not.

This code should really not ever be reached, so I might stand accused of
programming by superstition here, but if it is reached it would be good
to know about it.

Note: I'm not really sure how this menu is supposed to work in case
we're connected to an ad-hoc network where there is no AP, but that's a
problem for another day.

https://phabricator.endlessm.com/T4444