CodeView: launch GNOME Builder with the manifest of the application
Pass the manifest of the application to GNOME Builder to
construct the workbench for the application source. Concatenate
an override app-id that will be used to launch the modified app
so that there are no conflicts with the already running
application. The application itself needs to set the
G_APPLICATION_CAN_OVERRIDE_APP_ID flag in order for this to
work.