CodeView: listen to overview hiding and showing signals to hide overlay buttons
When the desktop is shown and the windows are hidden and vice
versa we need to make sure the overlay buttons are hidden and
shown again in those cases. The overview emits signals for
those cases used at other places in the shell already.

In the handlers we do not try to figure out the current active
session window and hide/show the existing buttons in any case.

https://phabricator.endlessm.com/T14811