Move eos-google-chrome-helper to Recommends
The code will now check for the existance of the GSettings schema
associated to this package, so it's no longer needed to make a
hard dependency on it to pass the tests. A 'Recommends' dependency
will be fine, while not breaking the build on ARM, where we don't
build eos-google-chrome-helper for.

https://phabricator.endlessm.com/T15084
https://phabricator.endlessm.com/T15084-debian