CodeView: Desktop hide/show, handle case where Builder and app are visible
The coding session consists of two windows, the GNOME Builder
one and the application one, each window has an overlay button
respectively. When we hide the Desktop we only want to show
the one that is associated with the currently focused window.

Follow-up of: cd6fa28429934e2b877967af6085a1d68326c148