iconGridLayout: add image default override option
By adding a JSON file in /var/eos-image-defaults/icon-grid,
the default icon grid for a locale can be overridden.
For instance, this allows the image builder to specify
having Chrome on the desktop by default rather than Chromium.

Follows the same syntax and locale-matching of the defaults in
/usr/share/eos-shell-content/icon-grid-defaults.

If there is any image default override that matches the user's locale,
it will take priority over any base default, even if the base default
is a more specific match to the user's locale.

Any image default prepending or appending is still applied
after this override.

https://phabricator.endlessm.com/T15324