updaterManager: Clear old notifications on updater state change
If the updater changes state, any notification which is currently being
shown by the shell is outdated (especially if it’s an error), so
explicitly clear it. This can happen even across multiple invocations of
eos-shell, since notifications are serialised by the message tray code.

This should fix the situation where a previous boot has shown an error
message about being unable to update, which gets re-shown on the next
boot even though the updater is happily in the Ready state now.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T15644