coding: Remove bouncing animation for button
Having this animation around results in two different animations -
one where we flip to a speedwagon window if the builder app
isnt launched and another where we bounce the button if it is
(or there is a speedwagon window for it already). This is confusing
so just remove the button animation.

An unfortunate side effect of this is that if builder is already
open then we don't have any feedback at all, though that can
be fixed later by allowing a speedwagon window to be shown
later.