Re-apply downstream translation changes
This includes the important clock string codes for Thai.

This also includes some customizations that we have made
in Spanish and Brazilian Portuguese, such as calling applications
"programas", using the informal "you", and a few other minor
wording changes.  As part of this, I cleaned up a few mistakes
in our downstream translations, such as the singular form being
a copy of the plural form, extra trailing spaces, and missing
periods.

https://phabricator.endlessm.com/T15801