util: return meaningful search engine when setting is not available
Now that the search entry at the desktop shows the search engine
name, we should make sure to return a meaningful value so that
the user really knows what's going on.

The current code, however, returns null both when the Chromium
settings are not found, or when the setting itself is not found
inside the settings file. In practice, this makes impossible to
check what's the cause of the empty string.

Fix that by checking the default browser and parsing the result
from the settings if available. If nothing valuable is found, just
return "Google", which is the default search engine anyway.

The string is translatable so we can ask translator to change the
provider name (e.g. to "Baidu" in China).

https://phabricator.endlessm.com/T15536