coding: Remove app/builder hiding on restack
The restack signal can be fired as a result of us changing the
active window during the flip animation. This will cause a window
to be hidden during the animation, which isn't any good.

I have not been able to reproduce the problem described in the
comments and besides, the builder/app window should be hidden
anyway at the end of the animation, so there's no need
to hide it again.