coding: Add _switchWindowsWithoutFlipping method
In some cases, we might need to switch the state of the session, but
flipping the window would just be the wrong thing to do. For instance,
playing a flip animation when the user unminimized the sibling window
or selected it from the overview just looks silly. In this case, we'll
want to call _switchWindowsWithoutFlipping, which just adjusts the state
and then shows and hides the relevant windows without doing anything
else.