shellEntry: change opacity on :hover and :focus
It is important to change the visual aspect of the search
bar when the user interacts with it, in order to demonstrate
that the search bar is actually interactive.

The current implementation, however, is static as it only
has one visual for all the states (focused, hovered and
normal).

Fix that by changing the background opacity on hover and focus.

https://phabricator.endlessm.com/T15536