Add a DBus method to add an app only if it's not visible
Sometimes we need to make sure an application is visible in the
desktop's grid (i.e. in its top level folder) but without forcing
an append of its icon because if the app is already visible it should
not move to the end of the grid.
For that reason a new AddAppIfNotVisible method is added to the DBus
interface which will only add the given app to the desktop if it's not
already there and visible.

https://phabricator.endlessm.com/T15941