Remove dependency on clutter and bump the dependency on mutter up to >= 3.22
In preparation for the rebase on GNOME Shell 3.22, we first want to update all
the underlying components without updating the shell, to follow an incremental
approach. The new mutter 3.22 is quite special, as it will place the shared
objects in a different path than older versions and will also ship its own
version of cogl and clutter, as bundled-in private libraries.

https://phabricator.endlessm.com/T16276