Moved from CDBS to debhelper, based on gnome-shell_3.22.3-2.debian.tar.xz
From this point on, we will be building and running eos-desktop on top of
mutter 3.22, which comes with its own private versions of clutter and cogl.

For that reason, we need to adapt the debian/rules so that eos-shell is
built taking into account the new layout of mutter's files and, instead
of just doing our own here, we just leverage this file from the gnome-shell
package we'll be rebasing our shell on top of next (gnome-shell_3.22.3-2).

https://phabricator.endlessm.com/T16276