Fix build errors due to deprecated functions and includes
The clutter-glx.h file is no longer shipped by mutter due to being
redundant with clutter-x11.h, so we need to remove it to build.

Functions cogl_clip_push_rectangle() and cogl_clip_pop() have been
deprecated since Clutter 1.16 and are no longer available in the
version of clutter shipped with mutter now (1.26.1), so we need to
replace it with equivalent implementations.

https://phabricator.endlessm.com/T16276