Don't prematurely add an icon coming from a folder to the main view
This is the reason why we have a hack in animateShuffling() to force
a repaint of the main view right before animating the icon, which we
can't use anymore as it relied in some undefined behaviour in clutter
that is no longer working as "expected", and now crashes the shell.

Besides, there is no need to artificially add the icon to the main view
when dragging it from inside a folder into the main grid layout, as it
will be inserted anyway in IconGridLayout.layout.repositionIcon.

Also, artificially adding that icon to the list really confused the logic
that finds out which icons need to be moved in the icon grid, which unmasked
a bug during the animation only visible after removing the call to paint(),
so this is seems to be also cleaner now.

https://phabricator.endlessm.com/T16707